separated functionality into different files, started datalinks
