+++ fornax/aws.py
@@ -1,81 +1,16 @@
-import requests
 import boto3
 import botocore
-from pathlib import Path
 import json
-import pyvo
 
-from astropy.utils.data import download_file
 from astropy.utils.console import ProgressBarOrSpinner
-from astropy.table import Table
-from astropy.io import votable
+from pathlib import Path
+
+from .prem import AccessPoint
 
 
-__all__ = ['AccessPoint', 'AWSAccessPoint', 'DataHandler']
+__all__ = ['AWSAccessPoint']
 
 
-
-class AccessPoint:
-    """A base class to handle a single data access point"""
-    
-    def __init__(self, url):
-        """An access point for on-prem server with direct url
-        
-        Parameters
-        ----------
-        url : the url to access the data
-        
-        """
-        
-        self.url  = url
-        self.type = 'prem'
-        self._accessible = None
-        
-    
-    def download(self, cache=True):
-        """Download data. Can be overloaded with different implimentation
-        
-        Parameters
-        ----------
-        cache : bool
-            Default is True. If file is found on disc it will not be downloaded again.
-            
-            
-        Return
-        ------
-        local_path : str
-            Returns the local path that the file was download to.
-            
-        """
-        
-        if self.url is None:
-            raise ValueError(f'No on-prem url has been defined.')
-        
-        log.info(f'downloading data from {self.type} using: {self.url}')
-        
-        path = download_file(self.url, cache=cache)
-        return path
-    
-
-    def is_accessible(self):
-        """Check if the url is accessible
-        
-        Return
-        ------
-        (accessible, msg) where accessible is a bool and msg is the failure message
-        
-        """
-        msg = ''
-        if self._accessible is None:
-            response = requests.head(self.url)
-            accessible = response.status_code == 200
-            if not accessible:
-                msg = response.reason
-            self._accessible = (accessible, msg)
-        return self._accessible
-    
-
-    
 class AWSAccessPoint(AccessPoint):
     """Handles a single access point on AWS"""
     
@@ -282,130 +217,3 @@ class AWSAccessPoint(AccessPoint):
 
             bkt.download_file(key, local_path, Callback=progress_callback)
         return local_path
-    
-    
-class AccessManager:
-    """AccessPoint container and manager"""
-    
-    def __init__(self, base_access):
-        """Initilize an AccessManager with a basic AccessPoint.
-        
-        Parameters
-        ----------
-        base_access: AccessPoint
-            a minimum access point with a simple url.
-        
-        """
-        
-        if not isinstance(base_access, AccessPoint):
-            raise ValueError(
-                f'type(base_access) is expected with be AccessApoint not {type(base_access)}'
-            )
-        
-        self.access_points = {base_access.type: [base_access]}
-        
-        # the default is the one to use. One of access_points
-        self.default_access_point = {base_access.type: base_access}
-        
-    
-    def add_access_point(self, access_point):
-        """Add a new AccessPoint to the manager
-        
-        Parameters:
-        -----------
-        access_point: AccessPoint, a subclass, or a list of them.
-            the access point to be added to the manager
-                
-        """
-        
-        # if a list, loop through the elements
-        if isinstance(access_point, list):
-            for ap in access_point:
-                self.add_access_point(ap)
-        else:
-            if not isinstance(access_point, AccessPoint):
-                raise ValueError(
-                    f'type(base_access) is expected with be AccessApoint, '
-                    f'a subclass or a list not {type(base_access)}'
-                )
-
-            ap_type = access_point.type
-            if not ap_type in self.access_points:
-                self.access_points[ap_type] = []
-            self.access_points[ap_type].append(access_point)
-        
-
-        
-    
-    
-    
-class DataHandler:
-    """A container for multiple AccessPoint instances"""
-    
-    def __init__(self, data_product, **kwargs):
-        """
-        Parameters
-        ----------
-        data_product: astropy.table or pyvo.dal.DALResults
-        
-        kwargs: keywrods arguments used to initialize the AccessPoint
-            instance or its subclasses.
-        
-        """
-        
-        if not isinstance(data_product, (pyvo.dal.DALResults, Table)):
-            raise ValueError(f'data_prodcut should be either '
-                              'astropy.table.Table or '
-                              'pyvo.dal.DALResults')
-        
-        # if we have an astropy table, convert to a pyvo.dal.DALResults
-        if isinstance(data_product, Table):
-            vot = votable.from_table(data_product)
-            dal_result = pyvo.dal.DALResults(vot)  
-        else:
-            dal_result = data_product
-        
-        
-        ## column name with direct access url 
-        # SIA v1
-        url_colname = dal_result.fieldname_with_ucd('VOX:Image_AccessReference')
-        if url_colname is None:
-            # SIA v2
-            if 'access_url' in dal_result.fieldnames:
-                url_colname = 'access_url'
-            else:
-                # try by ucd as a final resort
-                url_colname = dal_result.fieldname_with_ucd('meta.ref.url')
-        
-        # if still None, raise
-        # TODO allow the user the pass the name to avoid failing
-        if url_colname is None:
-            raise ValueError(f'Could not figure out the column with direct access url')
-        
-        # AccessPoint
-        # - self.access_points: some type of ap manager (could be a simple container; e.g. list)
-        # - each row has its access_points manager.
-        # - If we have a cloud_access column, use it. it is easier than datalinks as it does not
-        # require a new call to server.
-        # - elif we have datalinks, get all data links in one call, then pass on to AWS ap.
-        # - else: no aws info; fall back to on-prem
-        
-        # minimum access point that uses on-prem data
-        self.nrows = len(dal_result)
-        self.access_manager = [AccessManager(AccessPoint(url)) for url in dal_result[url_colname]]
-        
-        # if there is a 'cloud_access' json column, use it
-        if 'cloud_access' in dal_result.fieldnames:
-            profile = kwargs.get('profile', None)
-            
-            for irow in range(self.nrows):
-                jsontxt = dal_result[irow]['cloud_access']
-                awsAp = AWSAccessPoint.from_json(jsontxt, profile=profile)
-                self.access_manager[irow].add_access_point(awsAp)
-                
-        
-            
-            
-        from IPython import embed;embed();exit(0)
-        base_ap = [AccessPoint(url) for url in dal_result[url_colname]]
-        base_ap = AccessPoint()
